backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
python-dotenv==1.0.1
httpx==0.28.1
feedparser==6.0.11
python-slugify==8.0.4
supabase==2.10.0
langchain==0.3.13
langgraph==0.2.60
openai==1.57.2
pytest==8.3.4
